Duplicate content converted to functions
