+++ src/commands/server/create.ts
@@ -1,7 +1,5 @@
-import * as crypto from 'crypto';
 import { GluegunCommand } from 'gluegun';
 import { ExtendedGluegunToolbox } from '../../interfaces/extended-gluegun-toolbox';
-import * as fs from 'node:fs';
 
 /**
  * Create a new server
@@ -21,22 +19,23 @@ const NewCommand: GluegunCommand = {
       parameters,
       patching,
       print: { error, info, spin, success },
+      server,
       strings: { kebabCase },
       system,
       template,
     } = toolbox;
-
+    
     // Start timer
     const timer = system.startTimer();
-
+    
     // Info
     info('Create a new server');
-
+    
     // Check git
     if (!(await git.gitInstalled())) {
       return;
     }
-
+    
     // Get name
     const name = await helper.getInput(parameters.first, {
       name: 'server name',
@@ -45,17 +44,17 @@ const NewCommand: GluegunCommand = {
     if (!name) {
       return;
     }
-
+    
     // Set project directory
     const projectDir = kebabCase(name);
-
+    
     // Check if directory already exists
     if (filesystem.exists(projectDir)) {
       info(``);
       error(`There's already a folder named "${projectDir}" here.`);
       return undefined;
     }
-
+    
     // Clone git repository
     const cloneSpinner = spin('Clone https://github.com/lenneTech/nest-server-starter.git');
     await system.run(`git clone https://github.com/lenneTech/nest-server-starter.git ${projectDir}`);
@@ -63,67 +62,38 @@ const NewCommand: GluegunCommand = {
       filesystem.remove(`./${projectDir}/.git`);
       cloneSpinner.succeed('Repository cloned from https://github.com/lenneTech/nest-server-starter.git');
     }
-
+    
     // Check directory
     if (!filesystem.isDirectory(`./${projectDir}`)) {
       error(`The directory "${projectDir}" could not be created.`);
       return undefined;
     }
-
+    
     // Get description
     const description = await helper.getInput(parameters.second, {
       name: 'Description',
       showError: false,
     });
-
+    
     // Get author
     const author = await helper.getInput(parameters.second, {
       name: 'Author',
       showError: false,
     });
-
+    
     const prepareSpinner = spin('Prepare files');
-
+    
     // Set readme
     await template.generate({
       template: 'nest-server-starter/README.md.ejs',
       target: `./${projectDir}/README.md`,
       props: { name, description },
     });
-
- const configContent = fs.readFileSync(`./${projectDir}/src/config.env.ts`, 'utf8');
-
-    // Matches SECRET_OR_PRIVATE_KEY then any amount of anything until there is a '
-    const regex = /SECRET_OR_PRIVATE_KEY[^']*/g;
-
-    // if str aint defined its empty, when
-    const count = (str, pattern) => {
-      const re = new RegExp(pattern, 'gi')
-      return ((str || '').match(re) || []).length
-    }
-
-    const secretArr: string[] = []
-
-    for (let i = 0; i < count(configContent, regex); i++) {
-      secretArr.push(crypto.randomBytes(512).toString('base64'));
-    }
-
-    // Getting the config content and using native ts to replace the content because patching.update doest accept regex
-    let secretIndex = 0;
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const updatedContent = configContent.replace(regex, (match) => {
-      const secret = secretArr[secretIndex];
-      secretIndex++;
-      return secret;
-    });
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    await patching.update(`./${projectDir}/src/config.env.ts`, (content) => updatedContent);
-
-    await patching.update(`./${projectDir}/src/config.env.ts`, (data) =>
-      data.replace(/nest-server-/g, projectDir + '-')
-    );
-
+    
+    // Replace secret or private keys and remove `nest-server`
+    await patching.update(`./${projectDir}/src/config.env.ts`, (content) => server.replaceSecretOrPrivateKeys(content).replace(/nest-server-/g, projectDir
+      + '-'));
+    
     // Set package.json
     await patching.update(`./${projectDir}/package.json`, (config) => {
       config.author = author;
@@ -140,7 +110,7 @@ const NewCommand: GluegunCommand = {
       config.version = '0.0.1';
       return config;
     });
-
+    
     // Set package.json
     if (filesystem.exists(`./${projectDir}/src/meta`)) {
       await patching.update(`./${projectDir}/src/meta`, (config) => {
@@ -149,9 +119,9 @@ const NewCommand: GluegunCommand = {
         return config;
       });
     }
-
+    
     prepareSpinner.succeed('Files prepared');
-
+    
     // Init
     const installSpinner = spin('Install npm packages');
     await system.run(`cd ${projectDir} && npm i`);
@@ -163,7 +133,7 @@ const NewCommand: GluegunCommand = {
       );
       initGitSpinner.succeed('Git initialized');
     }
-
+    
     // We're done, so show what to do next
     info(``);
     success(
@@ -177,11 +147,11 @@ const NewCommand: GluegunCommand = {
     info(`  Run tests: npm run test:e2e`);
     info(`  Start server: npm start`);
     info(``);
-
+    
     if (!toolbox.parameters.options.fromGluegunMenu) {
       process.exit();
     }
-
+    
     // For tests
     return `new server ${name}`;
   },
